Fix: Split snapshot class in two.
